refactor: create seprate controller
